gauge-meter-plugin-develop/backend/requirements.txt
aiofiles==23.1.0
asyncio-mqtt==0.16.1
celery==5.2.7
cryptography==41.0.2
fastapi[all]>=0.95.0
hiredis==2.2.2
httpx==0.23.3
ilens-kafka-publisher==0.4.2.1
jinja2==3.1.2
jwt-signature-validator==0.0.5
kafka-python==2.0.2
lttb==0.3.1
matplotlib~=3.7.1
num2words==0.5.12
numpy~=1.23.5
orjson==3.8.9
paho-mqtt==1.6.1
pandas~=2.0.0
pendulum==2.1.2
pillow==9.5.0
pre-commit==3.2.2
prometheus-fastapi-instrumentator==6.0.0
psycopg2-binary==2.9.6
pycryptodome==3.17
pydantic==1.10.7
pyjwt==2.6.0
pymongo==4.3.3
pypika==0.48.9
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pyyaml~=6.0
redis==4.5.4
requests~=2.28.1
scipy==1.10.1
shortuuid==1.0.11
sqlalchemy==1.4.46
sqlparse==0.4.4
starlette~=0.27.0
ujson==5.7.0
uvicorn[standard]==0.21.1
word2number==1.1
XlsxWriter==3.0.9

